Create helper.ts and move files into it
